+++ ixmp/backend/base.py
@@ -445,6 +445,12 @@ class Backend(ABC):
         set_as_default
         """
 
+    def del_ts(self, ts: TimeSeries):
+        """OPTIONAL: Free memory associated with the TimeSeries *ts*.
+
+        The default implementation has no effect.
+        """
+
     @abstractmethod
     def check_out(self, ts: TimeSeries, timeseries_only):
         """Check out *ts* for modification.
@@ -1001,6 +1007,9 @@ class Backend(ABC):
 class CachingBackend(Backend):
     """Backend with additional features for caching data."""
 
+    #: :obj:`True` if caching is enabled.
+    cache_enabled = True
+
     #: Cache of values. Keys are given by :meth:`_cache_key`; values depend on
     #: the subclass' usage of the cache.
     _cache = {}
@@ -1009,12 +1018,22 @@ class CachingBackend(Backend):
     #: using :meth:`cache_get`.
     _cache_hit = {}
 
-    def __init__(self):
+    # Backend API methods
+
+    def __init__(self, cache_enabled=True):
         super().__init__()
+
+        self.cache_enabled = cache_enabled
 
         # Empty the cache
         self._cache = {}
         self._cache_hit = {}
+
+    def del_ts(self, ts: TimeSeries):
+        """Invalidate cache entries associated with *ts*."""
+        self.cache_invalidate(ts)
+
+    # New methods for CachingBackend
 
     @classmethod
     def _cache_key(self, ts, ix_type, name, filters=None):
@@ -1058,7 +1077,7 @@ class CachingBackend(Backend):
         """
         key = self._cache_key(ts, ix_type, name, filters)
 
-        if key in self._cache:
+        if self.cache_enabled and key in self._cache:
             self._cache_hit[key] = self._cache_hit.setdefault(key, 0) + 1
             return copy(self._cache[key])
         else:
@@ -1073,6 +1092,10 @@ class CachingBackend(Backend):
             :obj:`True` if the key was already in the cache and its value was
             overwritten.
         """
+        if not self.cache_enabled:
+            # Don't store anything if cache is disabled
+            return False
+
         key = self._cache_key(ts, ix_type, name, filters)
 
         refreshed = key in self._cache